+++ src/i18n/en_US/index.js
@@ -1,6 +1,24 @@
-// This is just an example,
-// so you can safely delete all default props below
-
 export default {
+  // library page
+  add: "Add",
+  info: "Detail Info",
+  // reader page
+  cursor: "Cursor",
+  highlight: "Hightlight",
+  comment: "Comment",
+  view: "View settings",
+  fullScreen: "Full Screen",
+  search: "Search Text",
+  rightMenu: "Right menu",
+  // settings page
+
+  // left menu
+  openedProjects: "Opened projects",
   library: "Library",
+  help: "Help",
+  settings: "Settings",
+  // project tree
+  // right menu
+  // meta tab
+  // annot card
 };
