+++ src/i18n/zh_CN/index.js
@@ -1,3 +1,23 @@
 export default {
-  library: "文档库",
+  // library page
+  add: "添加",
+  info: "详细信息",
+  // reader page
+  cursor: "光标",
+  highlight: "高亮",
+  comment: "注释",
+  view: "视图",
+  fullScreen: "全屏",
+  search: "搜索",
+  rightMenu: "右菜单",
+  // settings page
+
+  // left menu
+  openedProjects: "当前项目",
+  library: "项目库",
+  help: "帮助",
+  settings: "设置",
+  // right menu
+  // meta tab
+  // annot card
 };
