+++ src/pages/GLComponent.vue
