+++ src/pages/NotePage.vue
